moved node_text_length, children_counter to quantifiers.py
moved methods that 'quantify' a certain node's property or properties into quantifiers.py. This will hopefully help with moving the project towards a more functional, pipelinable direction
